Rename skycms_pl to gmina_trzebnica_pl and fix error handling
The SkyCMS API key identifies the municipality (tenant), not the user:
region ID 8 resolves to "Trzebnica 2" with this key but to "Brzezina"
with the key used by gmina_miekinia_pl. Scope the source and its docs to
Gmina Trzebnica accordingly.

- Fix SourceArgumentNotFound calls to pass (argument, value) as required
  by exceptions.py; they previously raised TypeError instead.
- Unknown region IDs return HTTP 200 with success=true and only a
  message, so the success check never fired and fetch() returned an
  empty schedule. Detect the missing garbage_kinds key and raise
  SourceArgumentNotFoundWithSuggestions listing the valid region IDs.
- Map the combined "Odpady Zielone / Kuchenne" stream to Icons.ORGANIC
  and make icon matching longest-key-first so it no longer depends on
  ICON_MAP insertion order.
- Use the repo-standard absolute imports and add SOURCE_CODEOWNERS.
